# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0",
        "description": "API Information",
        "contact": {
            "name": "Developer"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3001"
        }
    ],
    "paths": {
...
        "/dashboard/merchant": {
            "get": {
                "summary": "Get merchant dashboard details",
                "tags": [
                    "Dashboard"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "merchantId",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Merchant ID to retrieve dashboard details"
                    },
                    {
                        "in": "query",
                        "name": "start",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Start date for filtering transactions (optional)"
                    },
                    {
                        "in": "query",
                        "name": "end",
                        "schema": {
                            "type": "string"
                        },
                        "description": "End date for filtering transactions (optional)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant dashboard details retrieved successfully"
                    },
                    "400": {
                        "description": "Missing or invalid fields"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/dashboard/admin/": {
            "get": {
                "summary": "Get admin dashboard details",
                "tags": [
                    "Dashboard"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "start",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Start date for filtering transactions (optional)"
                    },
                    {
                        "in": "query",
                        "name": "end",
                        "schema": {
                            "type": "string"
                        },
                        "description": "End date for filtering transactions (optional)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin dashboard details retrieved successfully"
                    },
                    "400": {
                        "description": "Missing or invalid fields"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/merchant/": {
            "get": {
                "summary": "Get a list of all merchants",
                "tags": [
                    "Merchants"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of merchants retrieved successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "summary": "Update an existing merchant",
                "tags": [
                    "Merchants"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "Merchant data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "company_name": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "payment_volume": {
                                        "type": "number"
                                    },
                                    "commission": {
                                        "type": "number"
                                    },
                                    "settlementDuration": {
                                        "type": "number"
                                    },
                                    "jazzCashMerchantId": {
                                        "type": "string"
                                    },
                                    "easyPaisaMerchantId": {
                                        "type": "string"
                                    },
                                    "swichMerchantId": {
                                        "type": "string"
                                    },
                                    "webhook_url": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Merchant updated successfully"
                    },
                    "400": {
                        "description": "Missing or invalid fields"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "summary": "Add a new merchant",
                "tags": [
                    "Merchants"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "description": "New merchant data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "company_name": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "payment_volume": {
                                        "type": "number"
                                    },
                                    "commission": {
                                        "type": "number"
                                    },
                                    "settlementDuration": {
                                        "type": "number"
                                    },
                                    "jazzCashMerchantId": {
                                        "type": "string"
                                    },
                                    "easyPaisaMerchantId": {
                                        "type": "string"
                                    },
                                    "swichMerchantId": {
                                        "type": "string"
                                    },
                                    "webhook_url": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Merchant created successfully"
                    },
                    "400": {
                        "description": "Missing or invalid fields"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
...
}